# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Barbearia",
        "description": "Documentação dos endpoints da barbearia desenvolvida em Laravel 11",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local (Docker)"
        }
    ],
    "paths": {
        "/agendamentos": {
            "get": {
                "tags": [
                    "Agendamentos"
                ],
                "summary": "Listar toda a agenda da barbearia",
                "description": "Método responsável por listar toda a agenda de marcações da barbearia.",
                "operationId": "b235ba5d3e2d441a4ce2bc40a3856f8e",
                "responses": {
                    "200": {
                        "description": "Lista de agendamentos retornada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Agendamentos"
                ],
                "summary": "Criar um novo agendamento de horário",
                "description": "Método responsável por criar e reservar um novo agendamento de horário no sistema.",
                "operationId": "e3cb79c2253a67c2db2049f1813f0449",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cliente_id",
                                    "barbeiro_id",
                                    "servico_id",
                                    "data_hora"
                                ],
                                "properties": {
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "barbeiro_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "servicos": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            3
                                        ]
                                    },
                                    "data_hora": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-06-15 14:30:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Agendamento realizado com sucesso"
                    }
                }
            }
        },
        "/agendamentos/{id}": {
            "get": {
                "tags": [
                    "Agendamentos"
                ],
                "summary": "Consultar detalhes de um agendamento específico",
                "description": "Método responsável por consultar e exibir os detalhes específicos de um agendamento por meio de seu ID.",
                "operationId": "a531a083dec8a73c006b9316d54e5488",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do agendamento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do agendamento retornados"
                    },
                    "404": {
                        "description": "Agendamento não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Agendamentos"
                ],
                "summary": "Cancelar um agendamento",
                "description": "Método responsável por cancelar e remover permanentemente um horário agendado do sistema.",
                "operationId": "8a5b43b9c636fe4e496876453317d09a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do agendamento a ser cancelado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agendamento cancelado com sucesso"
                    },
                    "404": {
                        "description": "Agendamento não encontrado"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Agendamentos"
                ],
                "summary": "Remarcar data ou alterar dados do agendamento",
                "description": "Método responsável por remarcar a data/hora ou alterar o profissional associado a um agendamento existente.",
                "operationId": "a054c0963e464fff5cf489a609b65bd6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do agendamento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data_hora": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-06-15 16:00:00"
                                    },
                                    "barbeiro_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Horário reagendado com sucesso"
                    },
                    "404": {
                        "description": "Agendamento não encontrado"
                    }
                }
            }
        },
        "/agendamentos/disponibilidade": {
            "get": {
                "tags": [
                    "Agendamentos"
                ],
                "summary": "Consultar horários disponíveis de um barbeiro",
                "description": "Método responsável por calcular e retornar os horários disponíveis de um barbeiro em uma data específica.",
                "operationId": "09dd3e08ab7a70ee14d78436c315b6e8",
                "parameters": [
                    {
                        "name": "barbeiro_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "data",
                        "in": "query",
                        "description": "Formato YYYY-MM-DD",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "servicos[]",
                        "in": "query",
                        "description": "Array de IDs dos serviços ex: servicos[]=1&servicos[]=2",
                        "required": true,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de horários livres retornada"
                    }
                }
            }
        },
...
    },
...
}